+++ Changelog.md
@@ -1,4 +1,9 @@
-27/06/28201
+27/08/2021 
+
+* New function function LinkVB6Console(aFilename: string): Integer; in include vb6.pas
+
+
+27/06/2021
 
 * New Web Plugin
   * Add Function GetMIMEType( aURL : String) : string